refactor(rust): split electricity http client domains
